+++ CHANGES.md
@@ -3,6 +3,10 @@
 ### 0.1.4
 
 * `bijection-scrooge` for scrooge generated scala thrift code
+* Add tagged types, resolved Bijection debate
+* Implicit bijection between `List[T]`, `List[U]`
+* Implicit bijection between `Vector[T]`, `Vector[U]`
+* Implicit bijection between `Map[K1, V1]`, `Map[K2, V2]`
 
 ### 0.1.3